tests: mock namespace refactors
