+++ src/Remy.Tests/Integration/Console/DefaultRunnerTests.cs
@@ -6,6 +6,8 @@ using Remy.Console;
 using Remy.Console.Runners;
 using Remy.Core.Tasks;
 using Remy.Tests.StubsAndMocks;
+using Remy.Tests.StubsAndMocks.Core.Config.Yaml;
+using Remy.Tests.StubsAndMocks.Core.Tasks;
 using Serilog;
 
 namespace Remy.Tests.Integration.Console
