+++ src/Remy.Tests/Integration/TypeManagerTests.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using NUnit.Framework;
 using Remy.Core.Tasks;
 using Remy.Tests.StubsAndMocks;
+using Remy.Tests.StubsAndMocks.Core.Tasks;
 using Serilog;
 
 namespace Remy.Tests.Integration
